% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_stream.R
\name{agg_stream}
\alias{agg_stream}
\title{Aggregate segments}
\usage{
agg_stream(data, size = 100, value = "distance", windowed = TRUE)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{size}{split size}

\item{value}{time or distance of the size}

\item{windowed}{if it uses rolling window to generate segments.}
}
\value{

}
\description{
Aggregate segments
}
