% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_stream.R
\name{differential_stream}
\alias{differential_stream}
\title{Title}
\usage{
differential_stream(data)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}
}
\value{

}
\description{
Title
}
