% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev_correction.R
\name{ele_correction}
\alias{ele_correction}
\title{Title}
\usage{
ele_correction(data, z = 13, raster = NULL)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{z}{zoom level}

\item{raster}{raster object with elevations}
}
\value{

}
\description{
Title
}
