% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filelist_to_stream_df.R
\name{filelist_to_stream_df}
\alias{filelist_to_stream_df}
\title{Processes list of files containing GPS streams into a single dataframe}
\usage{
filelist_to_stream_df(filelist)
}
\arguments{
\item{filelist}{list of paths to the files containing the gpx or fit streams
file_to_stream_df()}
}
\description{
Processes list of files containing GPS streams into a single dataframe
}
