% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stream.R
\name{read_gpx}
\alias{read_gpx}
\title{Read gpx file}
\usage{
read_gpx(filename)
}
\arguments{
\item{filename}{The file path to the containing file
read_gpx()}
}
\description{
reads and parse gpx from file. supports coordinates, elevation and timestamp
}
