% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stream.R
\name{read_stream}
\alias{read_stream}
\title{Read GPS stream from file}
\usage{
read_stream(filename)
}
\arguments{
\item{filename}{The file path
read_stream()}
}
\value{
tibble with stream readings
}
\description{
Reads stream from a single gpx, kml, tcx or fit file and stores it in a Tibble.
Allows to process files from Strava's bulk export, which includes compressed fit
files (.fit.gz) and compressed kml files (.kmz)

Keeps the original variable names from the raw files
}
