% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stream_file.R
\name{read_stream_file}
\alias{read_stream_file}
\title{Load GPS streams from a single gpx or fit file and stores it in a Tibble
Allows to process files from Strava's bulk export, which includes compressed fit files (*.fit.gz)}
\usage{
read_stream_file(filename)
}
\arguments{
\item{filename}{The file path to the directory containing file
read_stream_file()}
}
\value{

}
\description{
Load GPS streams from a single gpx or fit file and stores it in a Tibble
Allows to process files from Strava's bulk export, which includes compressed fit files (*.fit.gz)
}
