% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_stream.R
\name{rolling_window}
\alias{rolling_window}
\title{Rolling window segment}
\usage{
rolling_window(data, size = 100, value = "distance")
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{size}{split size}

\item{value}{time or distance of the size}
}
\value{

}
\description{
Rolling window segment
}
