% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strava_metadata}
\alias{strava_metadata}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{achievement_count}{weight of the diamond, in carats}
\item{athlete.id}{weight of the diamond, in carats}
\item{athlete.resource_state}{weight of the diamond, in carats}
\item{athlete_count}{weight of the diamond, in carats}
\item{average_cadence}{weight of the diamond, in carats}
\item{average_heartrate}{weight of the diamond, in carats}
\item{average_speed}{weight of the diamond, in carats}
\item{average_temp}{weight of the diamond, in carats}
\item{average_watts}{weight of the diamond, in carats}
\item{comment_count}{weight of the diamond, in carats}
\item{commute}{weight of the diamond, in carats}
\item{device_watts}{weight of the diamond, in carats}
\item{display_hide_heartrate_option}{weight of the diamond, in carats}
\item{distance}{weight of the diamond, in carats}
\item{elapsed_time}{weight of the diamond, in carats}
\item{elev_high}{weight of the diamond, in carats}
\item{elev_low}{weight of the diamond, in carats}
\item{end_latlng1}{weight of the diamond, in carats}
\item{end_latlng2}{weight of the diamond, in carats}
\item{external_id}{weight of the diamond, in carats}
\item{flagged}{weight of the diamond, in carats}
\item{from_accepted_tag}{weight of the diamond, in carats}
\item{gear_id}{weight of the diamond, in carats}
\item{has_heartrate}{weight of the diamond, in carats}
\item{has_kudoed}{weight of the diamond, in carats}
\item{heartrate_opt_out}{weight of the diamond, in carats}
\item{id}{weight of the diamond, in carats}
\item{kilojoules}{weight of the diamond, in carats}
\item{kudos_count}{weight of the diamond, in carats}
\item{location_city}{weight of the diamond, in carats}
\item{location_country}{weight of the diamond, in carats}
\item{location_state}{weight of the diamond, in carats}
\item{manual}{weight of the diamond, in carats}
\item{map.id}{weight of the diamond, in carats}
\item{map.resource_state}{weight of the diamond, in carats}
\item{map.summary_polyline}{weight of the diamond, in carats}
\item{max_heartrate}{weight of the diamond, in carats}
\item{max_speed}{weight of the diamond, in carats}
\item{moving_time}{weight of the diamond, in carats}
\item{name}{weight of the diamond, in carats}
\item{photo_count}{weight of the diamond, in carats}
\item{pr_count}{weight of the diamond, in carats}
\item{private}{weight of the diamond, in carats}
\item{resource_state}{weight of the diamond, in carats}
\item{start_date}{weight of the diamond, in carats}
\item{start_date_local}{weight of the diamond, in carats}
\item{start_latitude}{weight of the diamond, in carats}
\item{start_latlng1}{weight of the diamond, in carats}
\item{start_latlng2}{weight of the diamond, in carats}
\item{start_longitude}{weight of the diamond, in carats}
\item{suffer_score}{weight of the diamond, in carats}
\item{timezone}{weight of the diamond, in carats}
\item{total_elevation_gain}{weight of the diamond, in carats}
\item{total_photo_count}{weight of the diamond, in carats}
\item{trainer}{weight of the diamond, in carats}
\item{type}{weight of the diamond, in carats}
\item{upload_id}{weight of the diamond, in carats}
\item{upload_id_str}{weight of the diamond, in carats}
\item{utc_offset}{weight of the diamond, in carats}
\item{visibility}{weight of the diamond, in carats}
\item{workout_type}{weight of the diamond, in carats}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
strava_metadata
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
