% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava_to_stream_df.R
\name{strava_to_stream_df}
\alias{strava_to_stream_df}
\title{Gathers GPS streams from Strava's API. returns a list containing the activity stream, and activity summary data}
\usage{
strava_to_stream_df(stoken, nacts = 1)
}
\arguments{
\item{stoken}{authorization token obtained from strava (see rStrava documentation)}

\item{nacts}{number of recent activities to get, default is 1, maximum value is 100
strava_to_stream_df()}
}
\description{
Gathers GPS streams from Strava's API. returns a list containing the activity stream, and activity summary data
}
