% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_stream.R
\name{summarise_stream}
\alias{summarise_stream}
\title{Summarise segment}
\usage{
summarise_stream(data)
}
\arguments{
\item{data}{dataframe or tibble containing a stream segment}
}
\description{
Summarise segment
}
