% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gpx.R
\name{write_gpx}
\alias{write_gpx}
\title{Write GPX file from data stream}
\usage{
write_gpx(stream, file = "file.gpx")
}
\arguments{
\item{stream}{data stream}

\item{file}{name of the file}
}
\value{

}
\description{
Write GPX file from data stream
}
