% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stream.R
\name{write_kml}
\alias{write_kml}
\title{write stream into KML file}
\usage{
write_kml(data, file = "file.kml")
}
\arguments{
\item{data}{data stream}

\item{file}{name of the file}
}
\value{

}
\description{
write stream into KML file
}
