% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_correction.R
\name{ele_correction}
\alias{ele_correction}
\title{Obtain elevation values from external source for stream dataframe}
\usage{
ele_correction(data, z = 13, raster = NULL, replace = TRUE)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{z}{zoom of Digital Elevation Model, between 6 and 14}

\item{raster}{raster object with DEM values}

\item{replace}{binary variable. Default value = TRUE. Defines whether to create a new column or replace the raw elevation}
}
\value{
GPS stream dataframe with new variable 'ele_DEM'
}
\description{
Uses SRTM data for obtaining elevation for each reading. Also supports external DEM in raster format.
If number of readings exceeds 200 points downloads a raster, else downloads each point at a time.
If local raster available, extract elevations directly
}
