% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev_correction.R
\name{ele_correction_point}
\alias{ele_correction_point}
\title{Title}
\usage{
ele_correction_point(data, z = 13)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{z}{zoom}
}
\value{

}
\description{
Title
}
