% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev_correction.R
\name{ele_correction_raster}
\alias{ele_correction_raster}
\title{Title}
\usage{
ele_correction_raster(data, z = 13)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{z}{zoom}
}
\value{

}
\description{
Title
}
