% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev_correction.R
\name{ele_correction_raster_local}
\alias{ele_correction_raster_local}
\title{Title}
\usage{
ele_correction_raster_local(data, raster, z = 13)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{raster}{raster object with elevations}

\item{z}{zoom}
}
\value{

}
\description{
Title
}
