% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stream.R
\name{match_stream}
\alias{match_stream}
\title{Compare 2 streams}
\usage{
match_stream(data, route, tolerance = 50)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{route}{another stream / route to be validated}

\item{tolerance}{threshold. Default setting on 20 meters}
}
\value{
either coordinates of match section(s) of data or character with explanation of mismatch
}
\description{
Evaluates if 2 streams belongs to the same route. Checks for distance to start and end of route, for completion of track
and for staying on the right track
}
