% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stream.R
\name{match_stream}
\alias{match_stream}
\title{Title}
\usage{
match_stream(data, route, tolerance = 20)
}
\arguments{
\item{data}{a stream}

\item{route}{another stream}

\item{tolerance}{thershold}
}
\value{

}
\description{
Title
}
