% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stream.R
\name{match_stream}
\alias{match_stream}
\title{Compare 2 streams}
\usage{
match_stream(data, route, tolerance = 20)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{route}{another stream / route to be validated}

\item{tolerance}{threshold. Default setting on 20 meters}
}
\value{

}
\description{
Evaluates if 2 streams belongs to the same route. Checks for spatial and temporal contiguity
}
