% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_stream.R
\name{rename_stream}
\alias{rename_stream}
\title{rename GPS streams according to convention long, lat, elev, timestamp and time}
\usage{
rename_stream(data)
}
\arguments{
\item{data}{a dataframe or tibble containing a GPS stream
rename_stream()}
}
\description{
rename GPS streams according to convention long, lat, elev, timestamp and time
}
