% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_stream.R
\name{rolling_window}
\alias{rolling_window}
\title{Title}
\usage{
rolling_window(data, largo)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{largo}{tamaño}
}
\value{

}
\description{
Title
}
