% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_stream.R
\name{smooth_stream}
\alias{smooth_stream}
\title{Generate model for lon, lat and ele in function of time. For stream without time, uses the GPS reading cumulative count.}
\usage{
smooth_stream(data, alpha = 0.05, interpolate = FALSE, replace = TRUE)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{alpha}{smoothing parameter}

\item{interpolate}{binary variable #¿para que?. Default setting on False}
}
\value{

}
\description{
Generate model for lon, lat and ele in function of time. For stream without time, uses the GPS reading cumulative count.
}
