% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_stream.R
\name{smooth_stream}
\alias{smooth_stream}
\title{Title}
\usage{
smooth_stream(data, alpha = 0.05, interpolate = FALSE)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{alpha}{smoothing parameter}

\item{interpolate}{binary}
}
\value{

}
\description{
Title
}
