% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_stream.R
\name{split_stream}
\alias{split_stream}
\title{Title}
\usage{
split_stream(data, size = 60, value = "time", windowed = FALSE)
}
\arguments{
\item{data}{dataframe or tibble containing a GPS stream}

\item{size}{tamaño splits}

\item{value}{tiempo o distancia del tamaño}

\item{windowed}{si ventanea o no}
}
\value{

}
\description{
Title
}
