% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strava_streams}
\alias{strava_streams}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{altitude}{weight of the diamond, in carats}
\item{cadence}{weight of the diamond, in carats}
\item{distance}{weight of the diamond, in carats}
\item{grade_smooth}{weight of the diamond, in carats}
\item{heartrate}{weight of the diamond, in carats}
\item{id}{weight of the diamond, in carats}
\item{lat}{weight of the diamond, in carats}
\item{lng}{weight of the diamond, in carats}
\item{moving}{weight of the diamond, in carats}
\item{temp}{weight of the diamond, in carats}
\item{time}{weight of the diamond, in carats}
\item{velocity_smooth}{weight of the diamond, in carats}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
strava_streams
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
