% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_stream.R
\name{uniform_stream}
\alias{uniform_stream}
\title{Uniform streams into name convention}
\usage{
uniform_stream(data)
}
\arguments{
\item{data}{is a dataframe or tibble containing a GPS stream. See read_stream documentation

uniform_stream()}
}
\value{
dataframe stream with coordinates and optionally elevation, time.
}
\description{
Rename streams according to convention lon (for longitude), lat (for latitude),
ele (for elevation or altitude), timestamp (for time signature of reading) and time (for seconds from start).
Other variables such as distance, power, cadence or heartrate are left as they were on original file
}
