% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stream.R
\name{write_gpx}
\alias{write_gpx}
\title{Write stream into GPX file}
\usage{
write_gpx(data, file = "file.gpx")
}
\arguments{
\item{data}{data stream}

\item{file}{name of the file}
}
\description{
Write stream into GPX file
}
